third_party/blink/web_tests/platform/generic/fast/mediacapturefromelement/CanvasCaptureMediaStream-webgl-events-expected.txt

Exercises the potential events on CanvasCaptureMediaStream.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Got a stream from canvas.
PASS stream.getVideoTracks().length is 1
PASS track.readyState is "live"
PASS Plugged stream to video tag.
PASS Video play callback succeeded.
PASS Drawing to canvas.
PASS Video canplaythrough callback succeeded.
PASS successfullyParsed is true

TEST COMPLETE
